Move component to Forms folder and fix TypeScript issue
Move ValidationErrorMessage to Forms folder. Update tsconfig file so that the parser can locate the component at its new location
