+++ tsconfig.json
@@ -36,7 +36,7 @@
     }
   },
   "include": [
-    "src"
+    "src", "src/Forms"
   ],
   "exclude": [
     "node_modules"
